backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.10
pydantic==2.8.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.1
